Remove util file and add reference for bottom sheet
